desktop_streamer_requirements.txt
PyGObject==3.44.1
psutil==5.9.6
PyYAML==6.0.1 